components/form: field select table - rename + validation
