+++ src/components/global/FormFieldTestWrapper.vue
@@ -16,7 +16,7 @@
  *
  * @components
  * - `FormFieldCompany`: Component to render company input field.
- * - `FormFieldCompanySelect`: Component to render company select field.
+ * - `FormFieldSelectTable`: Component to render company select field.
  * - `FormFieldEmail`: Component to render email input field.
  * - `FormFieldOptionGroup`: Component to render option group input field.
  * - `FormFieldPassword`: Component to render password input field.
@@ -35,7 +35,7 @@ import { defineComponent, ref } from 'vue';
 
 // components
 import FormFieldCompany from './FormFieldCompany.vue';
-import FormFieldCompanySelect from '../form/FormFieldCompanySelect.vue';
+import FormFieldSelectTable from '../form/FormFieldSelectTable.vue';
 import FormFieldEmail from './FormFieldEmail.vue';
 import FormFieldOptionGroup from '../form/FormFieldOptionGroup.vue';
 import FormFieldPassword from './FormFieldPassword.vue';
@@ -49,7 +49,7 @@ export default defineComponent({
   name: 'FormFieldTestWrapper',
   components: {
     FormFieldCompany,
-    FormFieldCompanySelect,
+    FormFieldSelectTable,
     FormFieldEmail,
     FormFieldOptionGroup,
     FormFieldPassword,
